# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
servers:
  - url: http://localhost:3000
security:
  - idAuth: []
...
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MethodNotAllowed:
      description: Method Not Allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    idAuth:
      description: |-
        this autentication method accepts a username like “Maria” without any password. If the username already exists, the user is logged in. If the username is new, the user is registered and logged in. The API will return the user identifier you need to pass into the Authorization header in any other API.
      type: http
      scheme: bearer      
  parameters:
...
    like_id:
      name: like_id
      in: path
      description: The like id. it is equivalent to the user_id of the person who likes it.
      required: true
      schema:
        $ref: "#/components/schemas/User_id/properties/user_id"
...
    user_id:
      name: user_id
      in: path
      description: The user id
      required: true
      schema:
        $ref: "#/components/schemas/User_id/properties/user_id"
    photo_id:
      name: photo_id
      in: path    
      description: the ID of the photo
      required: true
      schema:
        $ref: "#/components/schemas/Photo/properties/photo_id"
...
  schemas:
    Error:
      description: This object represent an error
      type: object
      properties:
        code:
           description: The Code error
           type: string
        message:
          description: The message error
          type: string
      required:
         - code
         - message
...
    Image:
      description: This object represent an image
      type: object
      properties:
        image_data:
          description: data of the Image
          type: string
          minLength: 10
          maxLength: 999999
          format: binary
          example: "010101110101"
      required:
        - image_data
    User_id:
      description: This object represent an user ID
      type: object
      properties:
        user_id:
          description: The user's ID
          type: integer
          format: int64
          minimum: 0
          maximum: 9999
          readOnly: true
          example: 20
      required:
        - user_id
...
    Photo:
      description: This object represent a photo
      type: object
      properties:
        datetime:
          description: The timestamp in ISO8601 format
          type: string
          format: date-time
          example: 2023-10-07T14:30:10
        likes:
          description: The photo's likes. it's an array of user_Id
          type: array
          minItems: 0
          maxItems: 9999
          items:
            $ref: "#/components/schemas/User_id"
        comments:
          description: the photo's comments
          type: array
          minItems: 0
          maxItems: 9999
          items:
            $ref: "#/components/schemas/Comment"
        photo_id:
          description: Photo ID
          type: integer
          format: int64
          minimum: 0
          maximum: 9999
          readOnly: true
          example: 2032
        author_id:          
          $ref: "#/components/schemas/User_id/properties/user_id"
      required:
        - datetime
        - likes
        - comments
        - photo_id
        - author_id
...
    Comment:
      description: This object represent a comment
      type: object
      properties:
        text:
          description: The content of the comment
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^.*?$'
          example: nice picture        
        author_id:          
          $ref: "#/components/schemas/User_id/properties/user_id"
        comment_id:
          description: The comment's ID
          type: integer
          format: int64
          minimum: 0
          maximum: 9999
          readOnly: true
          example: 2030
        photo_id:        
          $ref: "#/components/schemas/Photo/properties/photo_id"
        datetime:
          $ref: "#/components/schemas/Photo/properties/datetime"
      required:
        - text
        - photo_id
        - author_id
        - comment_id
        - datetime
...
paths:
...
  /users/{user_id}/photos: 
    parameters:
      - $ref: "#/components/parameters/user_id"               
    post:
      tags: ["photo"]
      operationId: uploadPhoto
      summary: Add a photo
      description: Add a photo using the object in the requestBody
      requestBody:
        description: An Image
        content:
          image/png:
            schema:
              $ref: "#/components/schemas/Image"            
        required: true
      responses:
        201:
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        405:
          $ref: '#/components/responses/MethodNotAllowed'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{user_id}/photos/{photo_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/photo_id"
    get:
      tags: ["photo"]
      operationId: getPhoto
      summary: Get an user photo
      description: |-
        this operation allows the user to obtain an user's photo.
      responses:
        200:
          description: photo get successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        401:
          $ref: '#/components/responses/Unauthorized'
        405:
          $ref: '#/components/responses/MethodNotAllowed'
        500:
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: ["photo"]
      operationId: deletePhoto
      summary: Delete a photo 
      description: |-
        Delete a user's photo. the photo can only be removed 
        if the author of that photo is the user himself
      responses:
        204:
          description: "Photo deleted successfully"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        405:
          $ref: '#/components/responses/MethodNotAllowed'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{user_id}/photos/{photo_id}/likes/{like_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/photo_id"
      - $ref: "#/components/parameters/like_id"
    put:
      tags: ["likes"]
      operationId: likePhoto
      summary: Add a Like
      description: Add a Like to a user's photo.     
      responses:
        204:
          description: Like added successfully
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        405:
          $ref: '#/components/responses/MethodNotAllowed'
        500:
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: ["likes"]
      operationId: unlikePhoto
      summary: remove a like
      description: |-
        remove a like from a user's photo.
      responses:
        204:
          description: Like removed successfully
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/MethodNotAllowed'
...